requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2b1
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-filter==2.4.0
django-ranged-response==0.2.0
django-redis==4.11.0
django-redis-sessions==0.6.2
django-simple-captcha==0.5.13
django-timezone-field==4.1.1
django-tinymce==2.7.0
djangorestframework==3.12.2
flower==0.9.7
generics==3.3.0
humanize==3.2.0
idna==2.10
importlib-metadata==3.7.0
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
Pillow==8.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.16
pycparser==2.20
PyMySQL==1.0.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tornado==6.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.0
